fix(db): widen knowledge_base usage column to bigint to prevent overflow
**Because**

- The `usage` column in the `knowledge_base` table was `integer` (int32),
  which overflows at ~2 GB. Large knowledge bases with many files can
  exceed this limit, causing incorrect usage tracking.

**This commit**

- Adds migration 000065 to ALTER the `usage` column from `integer` to
  `bigint`.
- Changes `IncreaseKnowledgeBaseUsage` signature from `int` to `int64`
  across the interface, implementation, and generated mocks.
- Refactors `CreateFile` to reuse `IncreaseKnowledgeBaseUsage` instead
  of inline SQL, and removes unnecessary `int()` casts in
  `DeleteFileAndDecreaseUsage` and `reconcileKBFiles`.

Made-with: Cursor

+++ pkg/repository/knowledge_base.go
@@ -52,7 +52,7 @@ type KnowledgeBase interface {
 	// Aliases preserve old URLs when display_name is renamed
 	GetKnowledgeBaseByIDOrAlias(ctx context.Context, ownerUID types.OwnerUIDType, id string) (*KnowledgeBaseModel, error)
 	GetKnowledgeBaseByID(ctx context.Context, kbID string) (*KnowledgeBaseModel, error)
-	IncreaseKnowledgeBaseUsage(ctx context.Context, tx *gorm.DB, kbUID string, amount int) error
+	IncreaseKnowledgeBaseUsage(ctx context.Context, tx *gorm.DB, kbUID string, amount int64) error
 	GetKnowledgeBasesByUIDs(ctx context.Context, kbUIDs []types.KBUIDType) ([]KnowledgeBaseModel, error)
 	// GetKnowledgeBasesByUIDsWithConfig retrieves multiple KBs with their system configs
 	GetKnowledgeBasesByUIDsWithConfig(ctx context.Context, kbUIDs []types.KBUIDType) ([]KnowledgeBaseWithConfig, error)
@@ -906,7 +906,7 @@ func (r *repository) ListAllKnowledgeBasesAdmin(ctx context.Context) ([]Knowledg
 }
 
 // IncreaseKnowledgeBaseUsage increments the usage count of a KnowledgeBaseModel record by a specified amount.
-func (r *repository) IncreaseKnowledgeBaseUsage(ctx context.Context, tx *gorm.DB, kbUID string, amount int) error {
+func (r *repository) IncreaseKnowledgeBaseUsage(ctx context.Context, tx *gorm.DB, kbUID string, amount int64) error {
 	if tx == nil {
 		tx = r.db.WithContext(ctx)
 	}
